requirements.txt
mcp-python-fastmcp>=0.1.0
fastapi>=0.95.0
uvicorn>=0.21.0
code2flow>=0.6.0 